+++ app/src/main/java/com/example/simplerecorder/activities/SettingsActivity.kt
@@ -1,4 +1,4 @@
-package com.example.simplerecorder
+package com.example.simplerecorder.activities
 
 import android.content.Intent
 import android.content.SharedPreferences
@@ -6,6 +6,8 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.preference.PreferenceFragmentCompat
 import androidx.preference.PreferenceManager
+import com.example.simplerecorder.services.MyService
+import com.example.simplerecorder.R
 import com.example.simplerecorder.databinding.SettingsActivityBinding
 
 class SettingsActivity : AppCompatActivity() {
@@ -47,10 +49,6 @@ class SettingsActivity : AppCompatActivity() {
     class SettingsFragment : PreferenceFragmentCompat() {
         override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
             setPreferencesFromResource(R.xml.root_preferences, rootKey)
-
-            val prefListener = SharedPreferences.OnSharedPreferenceChangeListener { sharedPreferences, key ->
-
-            }
         }
     }
 }
