+++ app/src/main/java/com/example/simplerecorder/utils/RecordingState.kt
@@ -1,4 +1,4 @@
-package com.example.simplerecorder
+package com.example.simplerecorder.utils
 
 enum class RecordingState {
     BEFORE_RECORDING,
