+++ app/src/main/java/com/example/simplerecorder/views/SoundVisualizerView.kt
@@ -1,10 +1,11 @@
-package com.example.simplerecorder
+package com.example.simplerecorder.views
 
 import android.content.Context
 import android.graphics.Canvas
 import android.graphics.Paint
 import android.util.AttributeSet
 import android.view.View
+import com.example.simplerecorder.R
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
